requirements.txt
tkinterdnd2
Pillow
